llm-receipt-parser/requirements.txt
# Core dependencies
openai
pillow

# Optional: for Claude alternative
# anthropic
